# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vinmo API",
        "description": "Documentação da API do Vinmo",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Lista eventos",
                "operationId": "e529c81e6aaec82f49950a87ed881f57",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Itens por página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Termo de busca",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "with_contract",
                        "in": "query",
                        "description": "Carrega o contrato relacionado",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Eventos retornados"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Events"
                ],
                "summary": "Cria um novo evento",
                "operationId": "e958e1c6c275d5f9b7ebcd837903d0ef",
                "responses": {
                    "200": {
                        "description": "Evento criado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/events/{event}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Exibe os detalhes de um evento",
                "operationId": "2f5e4d4d8d7989ca1f72be9c20d9daac",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "ID do evento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do evento"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Events"
                ],
                "summary": "Atualiza um evento",
                "operationId": "0c0cf4026635d252353fad3e5b48591c",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "ID do evento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Evento atualizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Events"
                ],
                "summary": "Remove um evento",
                "operationId": "b9043fb311444b3c3bcb101e1a643a9f",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "ID do evento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Evento removido"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/events/types/{contract}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Retorna os tipos de evento permitidos para um contrato",
                "operationId": "c46b71d2e6c3179de743596dd0043d52",
                "parameters": [
                    {
                        "name": "contract",
                        "in": "path",
                        "description": "ID do contrato",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tipos de evento retornados"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/events/{event}/images": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Retorna as imagens de um evento",
                "operationId": "510c80a3da181fcc15985f736ca258a7",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "ID do evento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Imagens retornadas"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/events/photos": {
            "post": {
                "tags": [
                    "EventPhotos"
                ],
                "summary": "Faz o upload de uma foto para o evento",
                "operationId": "9a3ee6c95e56dae0639e7b7053aa5877",
                "responses": {
                    "200": {
                        "description": "Foto salva com sucesso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}